+++ packages/css_colors/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.1.5
 
+* Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
+
+## 1.1.5
+
 * Updates minimum supported SDK version to Flutter 3.13/Dart 3.1.
 * Moves flutter_test and test dependencies to dev_dependencies.